Initialise Crunchyroll intro skipper
This commit adds the basic files required to stand up a functioning
Chrome extension to skip anime intros. The background script initially
checks to see if there are any comments on the page that specifies a
timestamp in which the intro ends, if found, it returns this. Otherwise,
we fall back to advertisement positions to skip intros.

A button is displayed on the player UI which allows the user to decide
whether they want to skip the intro, or not.