Add helper hint to skip button
This commit will add more information within the `skip intro` button to
allow the user to determine how the extension is extracting the
introduction end timestamp.

This commit also allows timestamp only comments to be used, ie. if a
commenter posted only a timestamp such as "1:21". We make the assumption
that the commenter's timestamp relates to when the into ends.